pyproject.toml
[project]
name = "lark-oapi-compact"
version = "0.1.3"
description = "Lark OpenAPI SDK for Python"
authors = [{ name = "L8ng", email = "straydragonl@foxmail.com" }]
readme = "README.md"
requires-python = ">=3.7"
license = { text = "MIT" }
keywords = [
  "feishu",
  "lark",
]
# https://pypi.org/classifiers/
classifiers = [
  "Development Status :: 4 - Beta",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3 :: Only",
  "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    # Upstream/v2_main https://github.com/larksuite/oapi-sdk-python
    "lark-oapi==1.3.4",
    # ---
    "pydantic==1.10.18",
    # Upstream/main old sdkhttps://github.com/larksuite/oapi-sdk-python
    "attrs>=23.2.0",
    "pycryptodome>=3.20.0",
    "requests>=2.31.0",
    "typing_extensions",
]

[project.urls]
Repository = "https://github.com/StrayDragon/oapi-sdk-python-compact"
Documentation = "https://github.com/StrayDragon/oapi-sdk-python-compact/blob/main/README.md"
Changelog = "https://github.com/StrayDragon/oapi-sdk-python-compact/blob/main/CHANGELOG.md"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "ipython>=7.34.0",
    "pre-commit>=2.21.0",
    "pytest-cov>=4.1.0",
    "pytest>=7.4.4",
    "ruff>=0.6.8",
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
